requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.8
graphql-relay==3.2.0
mysqlclient==2.2.8
pillow==12.1.1
promise==2.3
pyotp==2.9.0
python-dateutil==2.9.0.post0
qrcode==8.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
